Move `gpt_config` parsing to parent class.
