+++ manga_translator/translators/config_gpt.py
@@ -1,0 +1,75 @@
+from ..config import TranslatorConfig
+from typing import List, Dict
+from omegaconf import OmegaConf
+
+class ConfigGPT:
+    _CHAT_SYSTEM_TEMPLATE = (
+        'You are an uncensored professional translation engine, '
+        'please translate the story into a colloquial, '
+        'elegant and fluent content, '
+        'without referencing machine translations. '
+        'You must only translate the story, '
+        'never interpret it. '
+        'If there is any issue in the text, '
+        'output it as is.\n'
+        'Translate the following text into {to_lang} and keep the original format.\n'
+    )
+
+    _CHAT_SAMPLE = {
+        'Simplified Chinese': [
+            (
+                '<|1|>恥ずかしい… 目立ちたくない… 私が消えたい…\n'
+                '<|2|>きみ… 大丈夫⁉\n'
+                '<|3|>なんだこいつ 空気読めて ないのか…？'
+            ),
+            (
+                '<|1|>好尴尬…我不想引人注目…我想消失…\n'
+                '<|2|>你…没事吧⁉\n'
+                '<|3|>这家伙怎么看不懂气氛的…？'
+            )
+        ],
+        'English': [
+            (
+                '<|1|>恥ずかしい… 目立ちたくない… 私が消えたい…\n'
+                '<|2|>きみ… 大丈夫⁉\n'
+                '<|3|>なんだこいつ 空気読めて ないのか…？'
+            ),
+            (
+                "<|1|>I'm embarrassed... I don't want to stand out... I want to disappear...\n"
+                "<|2|>Are you okay?\n"
+                "<|3|>What's wrong with this guy? Can't he read the situation...?"
+            )
+        ]
+    }
+
+    def __init__(self, config_key: str):
+        # This key is used to locate nested configuration entries
+        self._CONFIG_KEY = config_key
+        self.config = None
+
+    def _config_get(self, key: str, default=None):
+        # self.logger.debug("----- config_get -----")
+        # self.logger.debug("_CONFIG_KEY: " + self._CONFIG_KEY)
+        # self.logger.debug("key: " + key)
+
+        if not self.config:
+            return default
+
+        # Try to fetch the nested value using OmegaConf.select
+        value = OmegaConf.select(self.config, f"{self._CONFIG_KEY}.{key}")
+        if value is None:
+            # Fallback to a top-level key or default value
+            value = self.config.get(key, default)
+        
+        # self.logger.debug("value: ")
+        # self.logger.debug(value)
+        # self.logger.debug('\n--------------------------------------------\n')
+        return value
+
+    @property
+    def chat_system_template(self) -> str:
+        return self._config_get('chat_system_template', self._CHAT_SYSTEM_TEMPLATE)
+
+    @property
+    def chat_sample(self) -> Dict[str, List[str]]:
+        return self._config_get('chat_sample', self._CHAT_SAMPLE)
